requirements.txt
amqp
billiard
celery
certifi
chardet
idna
kombu
pytz
requests
urllib3
vine
beautifulsoup4
bs4
lxml
urllib3
python-dotenv
IMDbPY
psycopg2